voice_sales/requirements.txt
python-dotenv==1.1.0
livekit==1.0.8
livekit-agents[silero,turn-detector]~=1.0
livekit-agents[google]
livekit-plugins-noise-cancellation~=0.2
